app/requirements.txt
fastapi
uvicorn
psycopg2-binary
kafka-python
pymongo